requirements.txt
requests==2.31.0
telethon==1.34.0
pillow==10.2.0
regex_spm==1.0.0
